feat: accept args for staticcheck
staticcheck needs some massaging of the input files given to it by
pre-commit - while staticcheck can run on single files, it will complain
if there are missing dependencies that get resolved in another file in
the same package. Furthermore, if it is passed several files, it demands
that those files live in the same package.

hence the approach has been to parse the file arguments and transform
them into their respective packages, calling staticcheck on one package
at a time.

the simple original solution did not allow for arguments to staticcheck
to be passed via pre-commit's `args` option.

we iterate on the original solution by parsing the passed arguments,
separating possible command line options and file arguments.

resolves #11